requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.1.1
reportlab==4.2.5
